Fix link to startpage
So heres this story: Because I tried to bang this entire project out after spending the entire day before studying, there were quite a few issues with the site. Most notably, the lack of the clock because I forgot to use camelcase within the JS file. Anyways, everything should be rectified, so there we go!